Rename project to fnanz
