+++ README.md
@@ -1,2 +1,2 @@
-# back-billetera
+# fnanz
 Proyecto de Gestión de Gastos (Back)
