--- src/main/java/com/ahumadamob/billeteraapi/controller/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Controladores REST.
- */
-package com.ahumadamob.billeteraapi.controller;
