--- src/main/java/com/ahumadamob/billeteraapi/repository/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Repositorios de acceso a datos.
- */
-package com.ahumadamob.billeteraapi.repository;
