--- src/main/java/com/ahumadamob/billeteraapi/service/jpa/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Implementaciones JPA de servicios.
- */
-package com.ahumadamob.billeteraapi.service.jpa;
