--- src/main/java/com/ahumadamob/billeteraapi/service/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Servicios de negocio.
- */
-package com.ahumadamob.billeteraapi.service;
