+++ src/main/java/com/ahumadamob/fnanz/controller/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Controladores REST.
+ */
+package com.ahumadamob.fnanz.controller;
