+++ src/main/java/com/ahumadamob/fnanz/entity/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Entidades de dominio de la aplicación.
  */
-package com.ahumadamob.billeteraapi.entity;
+package com.ahumadamob.fnanz.entity;
