+++ src/main/java/com/ahumadamob/fnanz/repository/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Repositorios de acceso a datos.
+ */
+package com.ahumadamob.fnanz.repository;
