+++ src/main/java/com/ahumadamob/fnanz/service/jpa/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Implementaciones JPA de servicios.
+ */
+package com.ahumadamob.fnanz.service.jpa;
