+++ src/main/java/com/ahumadamob/fnanz/service/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Servicios de negocio.
+ */
+package com.ahumadamob.fnanz.service;
